# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: MustaFlow AI API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /projects/{id}/files:
    get:
      operationId: listProjectFiles
      tags: [projects]
      summary: List generated files for a project
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      responses:
        "200":
          description: List of files
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ProjectFileSummary"
    post:
      operationId: createProjectFile
      tags: [projects]
      summary: Create a new file in the project
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProjectFileCreate"
      responses:
        "201":
          description: Created file contents
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProjectFileContent"
        "409":
          description: File path already exists
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
  /projects/{id}/files/{fileId}:
    get:
      operationId: getProjectFile
      tags: [projects]
      summary: Get the contents of one generated file by id
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
        - { name: fileId, in: path, required: true, schema: { type: integer } }
      responses:
        "200":
          description: File contents
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProjectFileContent"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
    patch:
      operationId: updateProjectFile
      tags: [projects]
      summary: Update the contents of a generated file
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
        - { name: fileId, in: path, required: true, schema: { type: integer } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProjectFileUpdate"
      responses:
        "200":
          description: Updated file contents
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProjectFileContent"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
    delete:
      operationId: deleteProjectFile
      tags: [projects]
      summary: Delete a project file by id
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
        - { name: fileId, in: path, required: true, schema: { type: integer } }
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                type: object
                required: [deleted]
                properties:
                  deleted: { type: boolean }
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
  /projects/{id}/files/{fileId}/diagnostics:
    post:
      operationId: getProjectFileDiagnostics
      tags: [projects]
      summary: Run a syntax/type check on a single file in the project's container and return structured diagnostics
      description: >-
        Runs tsc / node --check / py_compile inside the project's live
        container (auto-selected from file extension) and returns parsed
        diagnostics ready to render as Monaco markers. Requires the
        container to be running; returns ok=false with an error otherwise.
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
        - { name: fileId, in: path, required: true, schema: { type: integer } }
      responses:
        "200":
          description: Diagnostics result (ok=false when the container is unavailable)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProjectFileDiagnostics"
        "404":
          description: File not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
  /projects/{id}/files/{fileId}/rename:
    patch:
      operationId: renameProjectFile
      tags: [projects]
      summary: Rename (move) a project file to a new path
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
        - { name: fileId, in: path, required: true, schema: { type: integer } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProjectFileRename"
      responses:
        "200":
          description: Renamed file
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProjectFileContent"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "409":
          description: A file with that path already exists
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
...
  /projects/{id}/files/search:
    get:
      operationId: searchProjectFiles
      tags: [projects]
      summary: Full-text search across all project file contents
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
        - name: q
          in: query
          required: true
          schema: { type: string, minLength: 1 }
      responses:
        "200":
          description: Search results (max 50)
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/FileSearchResult"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
...
  /projects/{id}/files/{fileId}/raw:
    get:
      operationId: getProjectFileRaw
      tags: [projects]
      summary: Get raw (plain text) content of a project file by id
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
        - { name: fileId, in: path, required: true, schema: { type: integer } }
      responses:
        "200":
          description: Raw file content as plain text
          content:
            text/plain:
              schema:
                type: string
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
...
components:
  schemas:
...
    ApiError:
      type: object
      properties:
        error: { type: string }
      required: [error]
...
    ProjectFileSummary:
      type: object
      required: [id, path, mimeType, size, updatedAt]
      properties:
        id: { type: integer }
        path: { type: string }
        mimeType: { type: string }
        size: { type: integer }
        updatedAt: { type: string, format: date-time }
    ProjectFileContent:
      type: object
      required: [id, path, mimeType, content, updatedAt]
      properties:
        id: { type: integer }
        path: { type: string }
        mimeType: { type: string }
        content: { type: string }
        updatedAt: { type: string, format: date-time }
    ProjectFileCreate:
      type: object
      required: [path]
      properties:
        path: { type: string }
        content: { type: string }
    ProjectFileUpdate:
      type: object
      required: [content]
      properties:
        content: { type: string }
    ProjectFileRename:
      type: object
      required: [path]
      properties:
        path: { type: string }
    ProjectFileDiagnostic:
      type: object
      required: [severity, message]
      properties:
        line: { type: integer, nullable: true }
        column: { type: integer, nullable: true }
        severity:
          type: string
          enum: [error, warning, info]
        message: { type: string }
        source: { type: string, nullable: true }
    ProjectFileDiagnostics:
      type: object
      required: [ok, tool, diagnostics]
      properties:
        ok: { type: boolean }
        tool: { type: string, description: "Diagnostic tool used (tsc, node, python, eslint)" }
        diagnostics:
          type: array
          items: { $ref: "#/components/schemas/ProjectFileDiagnostic" }
        error: { type: string, nullable: true }
...
    FileSearchResult:
      type: object
      required: [fileId, file, lineNumber, lineContent]
      properties:
        fileId: { type: integer }
        file: { type: string }
        lineNumber: { type: integer }
        lineContent: { type: string }
...